fix: address PR review — util split, edit account/flag gating
- split customerVendor util into separate customer.ts / vendor.ts; getVendorName
  no longer falls back to externalId and uses Title Case like getCustomerName
- gate the edit affordance on showUploadOptions and only fetch custom accounts
  when the flag is on (useIsEditableCustomTransaction)
- keep edits on their original account: pin the PATCH url to the transaction's
  sourceAccountId and make the account field read-only in edit mode
