+++ src/msw/api/businesses/[business-id]/custom-accounts/[custom-account-id]/transactions/record/recordedCustomTransaction.ts
@@ -5,7 +5,8 @@ import { CategorizationStatus } from '@schemas/bankTransactions/bankTransaction'
 import { TransactionSource } from '@schemas/bankTransactions/base'
 import type { Categorization, Classification } from '@schemas/categorization'
 import { type RecordCustomTransaction, RecordCustomTransactionSchema } from '@schemas/customAccounts'
-import { getCustomerName, getVendorName } from '@utils/customerVendor'
+import { getCustomerName } from '@utils/customer'
+import { getVendorName } from '@utils/vendor'
 
 import { customAccountStore } from '@msw/api/businesses/[business-id]/custom-accounts/store'
 import { customerStore } from '@msw/api/businesses/[business-id]/customers/store'
