+++ src/utils/bankTransactions/shared.ts
@@ -9,7 +9,8 @@ import type { CategoryUpdate } from '@schemas/bankTransactions/categoryUpdate'
 import type { CustomAccount } from '@schemas/customAccounts'
 import { makeTagKeyValueFromTag } from '@schemas/tag'
 import { getDefaultTaxCodeForBankTransaction } from '@utils/bankTransactions/taxCode'
-import { getCustomerName, getVendorName } from '@utils/customerVendor'
+import { getCustomerName } from '@utils/customer'
+import { getVendorName } from '@utils/vendor'
 import { BankTransactionSelectionVariant } from '@providers/BankTransactionsCategorizationStore/BankTransactionsCategorizationStoreProvider'
 import type { BankTransactionNonSuggestedMatchOption } from '@providers/BankTransactionsCategorizationStore/utils'
 import { convertApiCategorizationToCategoryOrSplitAsOption } from '@components/BankTransactionCategoryComboBox/utils'
