+++ src/utils/customer.ts
@@ -1,5 +1,4 @@
 import type { Customer } from '@schemas/customer'
-import type { Vendor } from '@schemas/vendor'
 
 export function getCustomerName(
   customer?: Customer | null,
@@ -7,12 +6,4 @@ export function getCustomerName(
   return customer?.individualName
     ?? customer?.companyName
     ?? 'Unknown Customer'
-}
-
-export function getVendorName(
-  vendor?: Vendor | null,
-) {
-  return vendor?.individualName
-    ?? vendor?.companyName
-    ?? 'Unknown Vendor'
 }
