+++ src/utils/vendor.ts
@@ -1,0 +1,9 @@
+import type { Vendor } from '@schemas/vendor'
+
+export function getVendorName(
+  vendor?: Vendor | null,
+) {
+  return vendor?.individualName
+    ?? vendor?.companyName
+    ?? 'Unknown Vendor'
+}
